Split dequeing related functionality from Queue to its own class
